continue refactoring, finish up InvoiceAppController and add lots of TODOs
