+++ source/color_theme.py
@@ -1,0 +1,1 @@
+# TODO: Put colors that are used by the GUI here so they can be easily changed and referenced
